Update tests for organization
