+++ backend/models/checkin.test.js
@@ -1,9 +1,7 @@
-const Checkin = require("../../models/answer.model");
-const dbHandler = require("../db-handler");
+const Checkin = require("./checkIn.model");
 
-// Required database setup and teardown
-beforeAll(async () => await dbHandler.connect());
-afterAll(async () => await dbHandler.closeDatabase());
+const { setupDB } = require("../setup-test");
+setupDB("checkin-model");
 
 describe("Checkin Model saves the correct values", () => {
   test("Save a model instance and then read from the db", async (done) => {
