+++ backend/models/projectTeamMemer.test.js
@@ -1,9 +1,7 @@
-const ProjectTeamMember = require("../../models/project.model");
-const dbHandler = require("../db-handler");
+const ProjectTeamMember = require("./projectTeamMember.model");
 
-// Required database setup and teardown
-beforeAll(async () => await dbHandler.connect());
-afterAll(async () => await dbHandler.closeDatabase());
+const { setupDB } = require("../setup-test");
+setupDB("projectTeamMember-model");
 
 // Please add and expand on this simple test.
 describe("ProjectTeamMember Model saves the correct values", () => {
