+++ backend/models/recurringEvent.test.js
@@ -1,14 +1,13 @@
-const Event = require("../../models/event.model");
-const dbHandler = require("../db-handler");
+const RecurringEvent = require("./recurringEvent.model");
 
-// Required database setup and teardown
-beforeAll(async () => await dbHandler.connect());
-afterAll(async () => await dbHandler.closeDatabase());
+const { setupDB } = require("../setup-test");
+setupDB("recurringEvent-model");
 
-describe("Event Model saves the correct values", () => {
+// Please add and expand on this simple test.
+describe("Question Model saves the correct values", () => {
   test("Save a model instance and then read from the db", async (done) => {
     const submittedData = {
-      name: "eventName",
+      name: "testRecurringEvent",
       location: {
         // should we include address here?
         city: "Los Angeles",
@@ -16,26 +15,24 @@ describe("Event Model saves the correct values", () => {
         country: "USA",
       },
       hacknight: "Online", // DTLA, Westside, South LA, Online
+      brigade: "Hack for LA",
       eventType: "Workshop", // Project Meeting, Orientation, Workshop
-      description: "A workshop to do stuff",
+      description: "A test instance",
       date: 1594023390039,
       startTime: 1594023390039, // start date and time of the event
       endTime: 1594023390039, // end date and time of the event
-      hours: 2, // length of the event in hours
+      hours: 1594023390039, // length of the event in hours
       createdDate: 1594023390039, // date/time event was created
       updatedDate: 1594023390039, // date/time event was last updated
-      checkInReady: false, // is the event open for check-ins?
-      owner: {
-        ownerId: 33, // id of user who created event
-      },
+      checkInReady: true, // is the event open for check-ins?
     };
 
-    await Event.create(submittedData);
-    const savedDataArray = await Event.find();
+    await RecurringEvent.create(submittedData);
+    const savedDataArray = await RecurringEvent.find();
     const savedData = savedDataArray[0];
     expect(savedData.name === submittedData.name);
-    expect(savedData.location.city === submittedData.location.city);
-    expect(savedData.startTime === submittedData.startTime);
+    expect(savedData.location.country === submittedData.location.country);
+    expect(savedData.description === submittedData.description);
     done();
   });
 });
